+++ spec/modl/modl4.rb
@@ -4,9 +4,7 @@ require 'modl'
 
 a = <<~XXXXXXXX
 
-*LOAD="http://modules.num.uk/1.txt";
-o=Tesco:Every Little Helps:[t=Customer Service:0845555555];
-## o=Tesco;
+ab=ISO 8879\:1986;
 
 XXXXXXXX